Cleanup header
